requirements.txt
# Core Deep Learning Framework
torch>=1.8.0
torchvision>=0.9.0

# Numerical Computing
numpy>=1.19.0

# Configuration
PyYAML>=5.4.0
easydict>=1.9

# Image Processing
opencv-python>=4.5.0
Pillow>=8.0.0
scikit-image>=0.18.0

# Data Science & Machine Learning
scikit-learn>=0.24.0

# Visualization
matplotlib>=3.3.0

# Utilities
tqdm>=4.60.0

